requirements.txt
absl-py==0.8.1
asn1crypto==0.22.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.3.0
Automat==0.0.0
beautifulsoup4==4.8.2
category-encoders==2.1.0
certifi==2019.11.28
cffi==1.10.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==1.8.1
cssselect==1.1.0
cv2-tools==2.3.4
cycler==0.10.0
decorator==4.4.2
Django==2.1.5
dlib==19.6.1
Flask==1.1.1
gast==0.3.2
grpcio==1.25.0
h5py==2.10.0
hyperlink==17.1.1
idna==2.9
ImageHash==4.0
importlib-metadata==1.3.0
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mkl-service==2.3.0
more-itertools==8.0.2
murmurhash==1.0.1
networkx==2.4
numpy==1.16.6
opencv-python==4.1.1.26
packaging==20.0
pandas==0.25.3
parsel==1.2.0
patsy==0.5.1
Pillow==6.1.0
pluggy==0.13.1
protobuf==3.11.0
py==1.8.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.18
PyDispatcher==2.0.5
pydot==1.4.1
pymongo==3.10.1
pyOpenSSL==17.0.0
pyparsing==2.4.5
PyQt5==5.13.0
PyQt5-sip==12.7.1
pyquery==1.4.1
pytest==5.3.2
python-constraint==1.4.0
python-dateutil==2.7.5
pytils==0.3
pytz==2019.3
PyWavelets==1.1.1
pywin32==220
PyYAML==5.1.2
queuelib==1.4.2
redis==3.4.1
requests==2.23.0
scikit-learn==0.22.1
scipy==1.2.1
Scrapy==1.3.3
selenium==3.141.0
service-identity==17.0.0
six==1.13.0
sklearn2==0.0.13
soupsieve==2.0
statsmodels==0.10.2
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
tools==0.1.9
Twisted==17.5.0
urllib3==1.25.7
utils==1.0.0
w3lib==1.17.0
wcwidth==0.1.8
Werkzeug==0.16.0
wincertstore==0.2
zipp==0.6.0
zope.interface==4.4.2
